Add driver for system controller
Currently, it is only useful to read the CPUID. However, as a result
of wanting to print the CPUID, I also had to improve the implementation
of printf(). Currently, it only supports %i and %p flags, and the %i
is non-standard because it behaves like %x. This is good enough for
debugging purposes, but I may have to clean it up in the future.

Signed-off-by: Joel Nider <joel@nider.org>